==> 01_prepare_libraries.R <==
# Additionally, make sure to have the following R libraries installed:
#   
# The Rstan package (warning, there are 2 steps to the installation: Configuring C++ toolchains, and then installation of Rstan)
# Rmarkdown
# Shiny
# tidyverse
# BRMS

# install.packages("rstan", repos = c("https://mc-stan.org/r-packages/", getOption("repos")))
# remove.packages("rstan")
# if (file.exists(".RData")) file.remove(".RData")


# Sys.setenv(DOWNLOAD_STATIC_LIBV8 = 1) # only necessary for Linux without the nodejs library / headers
# install.packages("rstan", repos = "https://cloud.r-project.org/", dependencies = TRUE)


# example(stan_model, package = "rstan", run.dontrun = TRUE)

library("rstan") # observe startup messages
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = TRUE)

# install.packages("rmarkdown")
library("rmarkdown")

# install.packages("shiny")
library("shiny")

library(tidyverse)

# install.packages("brms")
library("brms")

library(ggplot2)
library(mltools)
library(data.table)
library(caret)
#install.packages("tidybayes")
library("tidybayes")
library(magrittr)
library(dplyr)
library(purrr)
library(forcats)
library(tidyr)
library(modelr)
library(ggdist)
library(tidybayes)
library(ggplot2)
library(cowplot)
library(rstan)
library(brms)
library(ggrepel)
library(RColorBrewer)
library(gganimate)
library(posterior)
library(distributional)
library("bayesplot")
source("01_prepare_libraries.R")
library("caret")
library(tidyverse); set.seed(42)
library(ggpmisc)
library(magrittr)
library(dplyr)
library(purrr)
library(forcats)
library(tidyr)
library(modelr)
library(ggdist)
library(tidybayes)
library(ggplot2)
library(cowplot)
library(rstan)
library(brms)
library(ggrepel)
library(RColorBrewer)
library(gganimate)
library(posterior)
library(distributional)
# install.packages("fdrtool")              # Install & load fdrtool package
library("fdrtool")
#install.packages("data.table")             # Install & load data.table package
library("data.table")
library(knitr)
library(gridExtra)
library(GGally)
library(mice)
library(nlme)
library(lme4)
library(mnormt)
library(boot)
# install.packages("HLMdiag")
library(HLMdiag)
# install.packages("kableExtra")
library(kableExtra)
# install.packages("pander")
library(pander)
library(tidyverse)
# install.packages("sjPlot")
library(sjPlot)
library(tidyverse)
library(lme4)
library(ggsci)
# install.packages("see")
library(see)
library(cowplot)
theme_set(theme_classic()) # This sets the default ggplot theme
library("ggpubr")
# install.packages("merTools")
library(merTools)
#install.packages("shiny")
library("shiny")
# install.packages("shinythemes")
library("shinythemes")
library(webshot)
# install.packages("ciTools")
library("ciTools")
library("tibble")
library(glmnet)
